refactor(macos): extract HandlerContext free functions; align CEFBridge with renamed CopyBrowserHost
HandlerContext.swift was 515 lines after rebase on the parallel
HandlerContext+Cookies split that landed via #62. Move the free
error/response helpers and the HandlerError enum into
HandlerResponses.swift, bringing HandlerContext to 455.

CEFBridge.mm in closeBrowser/setHidden was still calling the
pre-split name CopyBrowserHost; the introspection split renamed
it to CEFBridgeCopyBrowserHost. Update both call sites so the
post-rebase tree compiles.
